🚚 Move : 공통 컴포넌트 파일 이동
