+++ src/app/drink/_components/DrinkDescription.tsx
@@ -3,7 +3,7 @@
 import LikeButton from '@/components/common/LikeButton';
 import { useAuthStore } from '@/store/authStore';
 
-import ShareButton from './ShareButton';
+import ShareButton from '../../../components/common/ShareButton';
 
 type DrinkDescriptionProps = {
   name: string;
